fixing the function when K is provided
The input value of K is not used, instead the default value of `phiprime.mu` is used to calculate K values for individuals. So either the user has to calculate and enter `phi prime.mu` by himself and K is deleted from the input parameters or another option would be: `phi prime.mu` is calculated within the function if K is provided and by default K is `NaN`. I implemented the latter.